Removing some decoder constraints and improving backtracking.
Decoders over `HasNames` now works on Foldables.
Improved the decoder type to require explicit backtracking, so that
when decoder errors are returned to the user they are for the current
parse branch, rather than the last possible branch which is likely to
share little in common with the command users were trying to run.